move the generator to avoid circular import
