+++ datalogistik/util.py
@@ -18,7 +18,6 @@ import json
 import os
 import pathlib
 import shutil
-import time
 import uuid
 from collections.abc import Mapping
 
@@ -26,10 +25,8 @@ import pyarrow as pa
 import pyarrow.fs
 import urllib3
 
-from . import config, tpc_info
+from . import config
 from .log import log
-from .table import Table
-from .tpc_builders import DBGen, DSDGen
 
 
 def getenv_bool(name: str, default: bool = False) -> bool:
@@ -264,59 +261,6 @@ def get_arrow_schema(input_schema):
 
     output_schema = pa.schema(field_list)
     return output_schema
-
-
-# Generate a dataset by calling one of the supported external generators
-# TODO: Generator output cannot be used as dataset output directly, because of the
-# trailing columns.
-def generate_dataset(dataset):
-    log.info(f"Generating {dataset.name} data to cache...")
-    gen_start = time.perf_counter()
-    # This naming assumes the scale factor is always peresent, which is true for TPC-H but possibly not all generated datasets
-    dataset_path = dataset.ensure_dataset_loc(new_hash=f"raw_{dataset.scale_factor}")
-    generators = {"tpc-h": DBGen, "tpc-ds": DSDGen}
-
-    # override the format, since we only know how to directly generate tpc-raw format
-    dataset.format = "tpc-raw"
-
-    try:
-        generator_class = generators[dataset.name]
-        # TODO: support executable_path as env var?
-        generator = generator_class(executable_path=config.get_gen_location())
-
-        dataset_path.mkdir(parents=True, exist_ok=True)
-        generator.create_dataset(
-            out_dir=dataset_path,
-            scale_factor=dataset.scale_factor,
-            partitions=config.get_thread_count(),
-        )
-
-        metadata_table_list = []
-        for table in tpc_info.tpc_table_names[dataset.name]:
-            metadata_table_list.append(
-                Table(
-                    table=table,
-                    # These will always be multi_file, so we should code that
-                    multi_file=True,
-                    # TODO: is this line necessary?
-                    # this schema is not inferred, but it does not have
-                    # the same structure of a user-specified schema either
-                    # "schema": schema_to_dict(dataset.schema),
-                )
-            )
-        dataset.tables = metadata_table_list
-
-        gen_time = time.perf_counter() - gen_start
-        log.info("Finished generating.")
-        log.debug(f"generation took {gen_time:0.2f} s")
-        dataset.write_metadata()
-
-    except Exception:
-        log.error("An error occurred during generation.")
-        clean_cache_dir(dataset_path)
-        raise
-
-    return dataset
 
 
 def compress(uncompressed_file_path, output_dir, compression):
